pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "pyghgaq"
version = "0.0.1"
authors = [
  { name="César Ordóñez", email="cesar.ordonez@unige.ch" },
]
description = "A package to calculate GHG emissions from aquatic ecosystems"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = ["numpy>=2.3.4", "PyYaml"]

[project.urls]
Homepage = "https://github.com/ceordonez/PyGHGaq"
Issues = "https://github.com/ceordonez/PyGHGaq/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pyghgaq = ["*.yml"]
